Command:get and :set implemented for three primitive types: uint, float, string
(from here on, lua scripts are assumed to be utf8)

+++ MeshProc/ExtractSubMesh.cpp
@@ -41,7 +41,7 @@ bool ExtractSubMesh::Invoke()
 		{
 			if (!vMap.contains(t[i]))
 			{
-				vMap[t[i]] = vMap.size();
+				vMap[t[i]] = static_cast<uint32_t>(vMap.size());
 			}
 		}
 	}
